Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 int main() {
-    int a,b;
+    float a,b;
     char ch;
-    scanf("%d %d",&a,&b);
+    scanf("%f %f",&a,&b);
     scanf(" %c",&ch);
     if(ch=='+'){
-        printf("%d",a+b);
+        printf("%f",a+b);
     }else if(ch=='-'){
-        printf("%d",a-b);
+        printf("%f",a-b);
     }else if(ch=='*'){
-        printf("%d",a*b);
+        printf("%f",a*b);
     }else if(ch=='/'){
-        printf("%d",a/b);
+        printf("%f",a/b);
     }
     return 0;
 }
